Added configuration for GPIO ports and DTMF digit encoding
