+++ sipserv.c
@@ -48,6 +48,7 @@ Lesser General Public License for more details.
 #include <errno.h>
 #include <pjsua-lib/pjsua.h>
 #include "pi-out.c"
+#include "sipserv-structs.h"
 
 
 
@@ -62,41 +63,6 @@ Lesser General Public License for more details.
 
 // define max supported dtmf settings
 #define MAX_DTMF_SETTINGS 16
-
-// struct for app dtmf settings
-struct dtmf_config { 
-	int id;
-	int active;
-	int processing_active;
-	char *description;
-    char *audio_response_file;
-	char *tts_intro;
-	char *tts_answer;
-	char *cmd;
-};
-
-// struct for app configuration settings
-struct app_config {
-    int ipv6;
-	char *sip_domain;
-    int port;
-	char *sip_user;
-	char *sip_password;
-	char *language;
-	int record_calls;
-	int silent_mode;
-	char *tts;
-	char *announcement_file;
-	char *CallCmd;
-	char *AfterMath;
-	char *log_file;
-    char *config_file;
-	struct dtmf_config dtmf_cfg[MAX_DTMF_SETTINGS];
-    short gpio_enable;
-    short gpio_port[4];
-    short interrupt_send_port;
-    short dtmf_encoding;
-} app_cfg;  
 
 // global holder vars for further app arguments
 char *tts_file = "play.wav";
